Renamed core to node-http
